+++ engine/brain/__init__.py
@@ -1,0 +1,4 @@
+from engine.brain.base import Brain, Message
+from engine.brain.deepseek_api import DeepSeekAPIBrain
+
+__all__ = ["Brain", "Message", "DeepSeekAPIBrain"]